backend/requirements.txt
fastapi
uvicorn
python-dotenv
anthropic
requests
pydantic
python-multipart
